+++ Changelog.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)//
 
+## [Unreleased: 4.5.0-rc.13]
+ - Fix-10226: data from select boxes not displaying
+ 
 ## 4.5.0-rc.12
 ### Changed
  - Updated @formio/js@5.2.0-rc.6